[Bug #161893968] Refactor code
