+++ app/api/v2/__init__.py
@@ -1,7 +1,9 @@
 from flask import Blueprint
 from flask_restful import Api, Resource
 
-from .views import SignUp, Login, Product, OneProduct, Sale, OneSale, UpdateUser, Signout
+from .views.user_views import SignUp, Login, UpdateUser, Signout, GetUsers
+from .views.product_views import Product, OneProduct
+from .views.sales_views import Sale, OneSale
 
 '''Creation of blue prints and routes for endpoints'''
 blprint2 = Blueprint('api2', __name__, url_prefix="/api/v2")
@@ -13,4 +15,5 @@ api.add_resource(OneProduct, "/products/<int:productId>")
 api.add_resource(Sale, "/sales")
 api.add_resource(OneSale, "/sales/<int:saleId>")
 api.add_resource(UpdateUser, "/users/<int:userId>")
+api.add_resource(GetUsers, "/users")
 api.add_resource(Signout, "/auth/logout")
