+++ app/api/v2/models/product_models.py
@@ -1,5 +1,5 @@
 import psycopg2
-from flask import jsonify, abort, make_response
+from flask import jsonify, abort
 import datetime
 
 from .db_models import Db
@@ -20,8 +20,10 @@ class Product_Model(Db):
         '''Method to save a product by appending it to existing
         products table'''
         self.cursor.execute(
-            "INSERT INTO products(title,category,price,quantity,minimum_stock,description, date) VALUES(%s,%s,%s,%s,%s,%s,%s)",
-            (self.data["title"], self.data["category"], self.data["price"], self.data["quantity"],
+            """INSERT INTO products(title,category,price,quantity,
+            minimum_stock,description, date) VALUES(%s,%s,%s,%s,%s,%s,%s)""",
+            (self.data["title"], self.data["category"], self.data["price"],
+             self.data["quantity"],
              self.data["minimum_stock"], self.data["description"], self.date),
         )
         self.cursor.execute("SELECT id FROM products WHERE title = %s",
@@ -29,30 +31,31 @@ class Product_Model(Db):
         row = self.cursor.fetchone()
         self.id = row[0]
 
-    def update(self, productId, title, category, price, quantity, minimum_stock, description):
+    def update(self, productId):
         '''Method is meant to update a product by editing its details in the
         products table'''
-        self.db = Db()
-        self.conn = self.db.createConnection()
-        self.db.createTables()
-        self.cursor = self.conn.cursor()
+        self.productId = productId
         self.cursor.execute("SELECT id FROM products WHERE title = %s",
-                            (title,))
+                            (self.data["title"],))
         row = self.cursor.fetchone()
         self.date = datetime.datetime.now()
-
-        if not row or row[0] == productId:
+        if not row or row[0] == self.productId:
             try:
                 self.cursor.execute(
-                    """UPDATE products SET title = '{}', category = '{}', price = '{}',
-                        quantity = '{}', minimum_stock = '{}', description = '{}', date='{}'
-                        WHERE id = {}
-                        """.format(title, category, int(price), int(quantity),
-                                   int(minimum_stock), description, self.date, productId)
+                    """UPDATE products SET title = %s, category = %s, price = %s,
+                        quantity = %s, minimum_stock = %s, description = %s,
+                         date = %s
+                        where title = %s
+                        """,
+                    (self.data["title"], self.data["category"],
+                     self.data["price"], self.data["quantity"],
+                     self.data["minimum_stock"], self.data["description"],
+                     self.date, self.data["title"])
                 )
-
             except Exception as e:
-                abort(403, "Error: Failed to update details")
+                print(e)
+        else:
+            abort(403, "Product title already exists, try another one")
 
     def get(self):
         sql = "SELECT * FROM products"
@@ -74,13 +77,10 @@ class Product_Model(Db):
 
     def delete(self, productId):
         self.productId = productId
-        try:
-            self.cursor.execute(
-                "DELETE from products where id = %s",
-                (self.productId,)
-            )
-        except Exception:
-            abort(403, "Error: Failed to delete details")
+        self.cursor.execute(
+            "DELETE from products where id = %s",
+            (self.productId,)
+        )
 
     def updateQuanitity(self, quantity, title):
         '''Method is meant to update a product by editing its details in the
