+++ app/api/v2/models/sale_models.py
@@ -8,7 +8,8 @@ from .db_models import Db
 class Sales_Model():
     '''Initializes a sale'''
 
-    def __init__(self, email=None, product=None, quantity=None, subtotals=None):
+    def __init__(self, email=None, product=None,
+                 quantity=None, subtotals=None):
         if email or product or quantity or subtotals:
             self.email = email
             self.title = product["title"]
@@ -23,8 +24,10 @@ class Sales_Model():
     def save(self):
         '''Saves a sale to sale records'''
         self.cursor.execute(
-            "INSERT INTO sales(email, title, quantity, subtotals, date) VALUES(%s,%s,%s,%s,%s)",
-            (self.email, self.title, self.quantity, self.subtotals, self.date),)
+            """INSERT INTO sales(email, title, quantity, subtotals,
+             date) VALUES(%s,%s,%s,%s,%s)""",
+            (self.email, self.title, self.quantity,
+             self.subtotals, self.date),)
 
     def get(self):
         self.cursor = self.conn.cursor()
@@ -42,9 +45,9 @@ class Sales_Model():
             onesale["subtotals"] = list_of_sales[4]
             onesale["date"] = list_of_sales[5]
             allsales.append(onesale)
-      
+
         return allsales
-    
+
     def checkSales(self):
         sale1 = Sales_Model()
         sales = sale1.get()
