+++ app/api/v2/models/user_model.py
@@ -22,7 +22,8 @@ class User_Model(Db):
             "INSERT INTO users(email,password,admin) VALUES(%s,%s,%s)", (
                 self.email, self.password, self.admin,)
         )
-        self.cursor.execute("SELECT id FROM users WHERE email = %s", (self.email,))
+        self.cursor.execute("SELECT id FROM users WHERE email = %s",
+                            (self.email,))
         row = self.cursor.fetchone()
         self.id = row[0]
 
@@ -45,10 +46,13 @@ class User_Model(Db):
         self.cursor.execute(
             "UPDATE users SET admin = %s WHERE id = %s", (True, userId,)
         )
-    
+
     def logout(self, token, date):
         '''method to logout a user from the system'''
         self.cursor.execute(
-                "INSERT INTO blacklist (token, date) VALUES (%s,%s)",
-                (token, date,)
+            "INSERT INTO blacklist (token, date) VALUES (%s,%s)",
+            (token, date,)
         )
+
+    def __repr__(self):
+        return self.conn.close()
