+++ app/test/v2/test_sales.py
@@ -78,38 +78,6 @@ class TestSales(TestsForApi):
                          "Must enter the product details in the body")
         self.assertEqual(resp.status_code, 400)
 
-    def test_post_sale_product_id_wrong_data_type(self):
-        '''Test to assert the correct datatype for product id when
-        making a sale'''
-        resp = self.test_client.post("/api/v2/sales",
-                                     data=json.dumps({
-                                         "title": 2901,
-                                         "quantity": 1
-                                     }),
-                                     headers={
-                                         'x-access-token': self.attendant_token,
-                                         'content-type': 'application/json'
-                                     })
-        response = json.loads(resp.data)
-        self.assertEqual(response["message"], "title must be a string")
-        self.assertEqual(resp.status_code, 400)
-
-    def test_post_sale_wrong_quantity_data_type(self):
-        '''Test to assert the correct datatype for quantity when
-        making a sale'''
-        resp = self.test_client.post("/api/v2/sales",
-                                     data=json.dumps({
-                                         "title": "orange",
-                                         "quantity": "one"
-                                     }),
-                                     headers={
-                                         'x-access-token': self.attendant_token,
-                                         'content-type': 'application/json'
-                                     })
-        response = json.loads(resp.data)
-        self.assertEqual(response["message"], "Quantity must be an integer")
-        self.assertEqual(resp.status_code, 400)
-
     def test_post_sale_many_fields_given(self):
         '''Test to assert the result if many fields are given'''
         resp = self.test_client.post("/api/v2/sales",
@@ -154,7 +122,7 @@ class TestSales(TestsForApi):
                                          'content-type': 'application/json'
                                      })
         response = json.loads(resp.data)
-        self.assertEqual(response["Message"], "Must be an attendant to undertake this action")
+        self.assertEqual(response["Message"], "Must be an attendant!")
         self.assertEqual(resp.status_code, 401)
 
     def test_get_all_sales_admin(self):
@@ -174,8 +142,8 @@ class TestSales(TestsForApi):
                                         'x-access-token': self.attendant_token
                                     })
         response = json.loads(resp.data)
-        self.assertEqual(response["Message"], "Must be an admin to undertake this action")
-        self.assertEqual(resp.status_code, 401)
+        self.assertEqual(response["Message"], "Success")
+        self.assertEqual(resp.status_code, 200)
 
     def test_get_one_sale_admin(self):
         '''Test for getting one sale for an admin'''
