+++ run.py
@@ -3,6 +3,7 @@ from instance.config import Config
 
 app = create_app(config_name="development")
 
+
 @app.route("/")
 def index():
     return "<p>Find the app documentation <a href='https://jemostoremanager.docs.apiary.io/'>here</a></p>"
